LPS-101448 Minor refactor for sortable items
